Fix from avva:
Attached is a small patch fixes a bug in talkpost_do.bml which occurs
in the following scenario:

1. Somebody leaves an anonymous comment in X's journal.
2. Y replies to that comment.
3. X receives mail with Y's comment where the quoted part is not the
anonymous comment replied to by Y, but rather X's original post.

The bug happened due to an overly ambitious SQL query which extracted
both parent comment body and parent comment username at once, and
failed when there was no parent comment username (since the parent
comment was anonymous), and so even the body was unavailable.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@355 1dfba125-df10-0410-8a77-af1a8d6d074f