[polls working & ampersand policy change] The main point of this patch was to make poll cooperate in clustered journals.
In the process I cleaned a lot of crap up and realized something: when
we make URLs to pass off to the style system, or make them in
functions like LJ::item_link, we shouldn't be making & into &amp;
because we *think* it'll be used in HTML.  We really can't know how
the style user will use it.  It's up to the style to do a
transformation on it if it's used in an attribute.  Like:
%%eh:readurl%%

Also, I removed the &nc= (number comments) crap, because I hate it.
Sorry, mart. (or whoever did like it)  It's really wrong.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@757 1dfba125-df10-0410-8a77-af1a8d6d074f