db & code changes to make the friends data table over 30% smaller. changing it (eventually) from a char(7)+char(7) to a mediumint+mediumint for the color info.  no reason to use 7 bytes per color when it's a 24 bit value.
so for now we bloat it up to 8 bytes per color, to store the 24 bit
color in decimal for awhile, then convert those 16 bytes does to 6.

this may not seem important now, but after cluster conversion is done,
and we can kill all the 43GB of crap on the master db, this is one of
the largest tables remaining, yet it's under 100 MB... it'll be down
to like 60 MB or so.  :)


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@866 1dfba125-df10-0410-8a77-af1a8d6d074f