avva slipped this by without me noticing.  :P
the user table uses only fixed-size fields for speed and reliability.
that is, no VARCHARs.  which means user record takes the full amount
of bytes possible, so i'm shrinking this from 100 to 80.  really,
we need to have a userprop for journal title, which can be really long.
i posted to lj_dev about this ages ago.  journal name and journal title
are different ideas.

even with CJK utf-8, 80 should be enough for a name... 80/3 = 26 chars.
my full name is 26 chars: "bradley joseph fitzpatrick".  we can change
this later if it's really a problem, but i don't think it will be.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@920 1dfba125-df10-0410-8a77-af1a8d6d074f