new BML function:  BML::decide_langage()
let's BML client code re-run the language decision logic, in case
it changed something (like the cookies).  without this, all the client
code could do is fall-back to something like the DefaultLanguage, ignoring
the Accept-Language entirely.

this also cleans up both BML processor's handler loop, since all the logic
is moved.

and since it's in a sub, and we can just return when we know the answer,
it's also cleaner logic, and perhaps even faster, but that wasn't a goal.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@1110 1dfba125-df10-0410-8a77-af1a8d6d074f