Excellent bug find by <lj user=avva> in the new friends view caching code.
He explains:

BF> But, we already send $lastmax off to LJ::get_recent_items in the form of
BF> the 'notafter' key.

BF> Is that not working?  Is the parameter off by one in the wrong direction?
BF> (I adjusted it by one to compensate for the <= the get_recent_items does)

No, that's what I thought at first, but after I rechecked I realised
it was in the right direction.

BF> This seems like another way to do something I thought it was already
BF> doing.

OK, an example. Let's say rlogtime of right now = 50. We want to build
a view with $getitems=25. We have a cache of 25
items with rlogtimes from 75 to 99 (99 is the oldest, 75 is the
youngest). After we load them into @items, we set $lastmax=74, because
we don't want to load anything older than rlogtime=74 (this is taken care of by
sending $lastmax as 'notafter' to get_recent_items() ).

Now we continue building, and get a friend with one new entry at
rlogtime=60. We load the entry, merge it into @items and throw away
the oldest entry there, so now @items looks like this:

60 75 76 77.... 98

And now we set $lastmax=$items[-1]->{'rlogtime'} , so now suddenly
$lastmax=98.

If we now load another friend who has a new entry at 65 and an old
entry at 80 (which is already in our cache), get_recent_items() will
load both because $lastmax is 98. And the old entry of rlogtime=80
will join the one already in @items, loaded from the cache, so now we
have a duplicate entry in @items.

Before the cache, the assignment $lastmax=$items[-1]->{'rlogtime'} always
_lowered_ $lastmax, and we used it to avoid loading obviously unnecessary
stuff. But now this assignment can raise it too high, if we already have
@items full of stuff from the cache that we don't want to load twice.

So my patch says: change $lastmax as before, but never raise it as
high as the cache items, so we don't load them again.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@1724 1dfba125-df10-0410-8a77-af1a8d6d074f