Deleted comments are visible to all users http://zilla.livejournal.org/show_bug.cgi?id=1133
------- Additional Comment #5 From Anatoly Vorobey  2003-07-07 02:17 -------

This bug happens due to combination of two unrelated reasons:
1. talkpost.bml doesn't check for "deleted" status of a parent comment.
2. When we delete a comment, its text is deleted in the DB, but not in memcache.

Previously, before memcache, only 1. was true, which made it possible to
sometimes look up the author of a deleted comment, but never the text (since it
was deleted in the DB). The current situation is much worse.

I'll have the fix for this ready shortly.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@2666 1dfba125-df10-0410-8a77-af1a8d6d074f