Make translation (ML) system support plural mapping http://zilla.livejournal.org/show_bug.cgi?id=1160
In addition to [[foo]] interpolations in translation strings, we now
also support conditional lookups.  You provide the variable name
which the caller should be providing as a number, then LJ::Lang
(which knows the current language) maps that to the correct plural
case for that language, and picks the correct word from the list.

So, before:

You have [[num]] entry/entries.

That's lame.

Now:

You have [[num]] [[?num|entry|entries]].

Say "num" is 2.

And BML/LJ knows the current language is en_LJ or en or en_GB ("en"), so it figures out
that 2 is plural in English (case 0 = singular, case 1 = plural).

So it takes the ith index of the list you give it, and index 1 is "entries".

So you get:

You have 2 entries.

We've provided count to case mappings for English, Russian, and
French.  Others we can add later easily.

No support for ordinal numbers yet, but that could be easy.  Maybe
like:  [[?num,ord]] or something.  *shrug*.

Props to <lj user=daveman692> for finishing and testing all this after
I typed some of it out and said, "eh, should work".  He's going to be
commiting more patches soon which start using this.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@3438 1dfba125-df10-0410-8a77-af1a8d6d074f