more paranoid moveucluster that only starts to move when it's sure it knows how to do everything.  (could save us from an accident in the future)
but also, in the process, discovered we're not moving userpropblob!  so fixed that.

(but I had a feeling there was something missing, but this was just an easier
way to find it than going through each table by hand and verifying, which
felt error-prone)

also added a flag to override the "already moving" flag, which isn't needed often,
but is often during test locally.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@3457 1dfba125-df10-0410-8a77-af1a8d6d074f