add fast paths to the html cleaner for both comment and journal entry cleaning.
in short, the html cleaner is now 2x-3x faster on posts/comments that
contain no markup and no URLs to auto-linkify because we don't have to
make a new HTML::Parser object, etc.

the html cleaner has always been the big CPU suck, so this should help a lot.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@3501 1dfba125-df10-0410-8a77-af1a8d6d074f