http://zilla.livejournal.org/1689 (part 1)
This commit puts in the reluser2 code that is going to be used for having
clustered user edges.  This code provides all of the framework, but nothing
uses it just yet.

All of the reluser functions (LJ::get_rel_user, etc) have been overloaded
to support reluser and reluser2 depending on what type you send to them.
Single letter types such as 'B' and 'A' get directed to the non-clustered
"reluser" table, but word types such as 'hide_comm_assoc' get directed to
the clustered reluser2.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@3693 1dfba125-df10-0410-8a77-af1a8d6d074f