http://zilla.livejournal.org/1777
Add a feature to support so that they can grant the "supportchangesummary"
priv.  This priv requires supportmovetouch and supportmakeinternal to
function, because those all work together.

This patch also made a lot of the priv checking code much cleaner.  It used
to repeat the same block of code 4 times, so I made it a helper function.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@3780 1dfba125-df10-0410-8a77-af1a8d6d074f