http://zilla.livejournal.org/1781
New feature: spam reporting system.  When a comment is deleted, users will now
have the option to "mark comment as spam."  When this is checked, the comment
is saved for the admins to review it later.

An admin tool (/admin/spamreports.bml hinging on siteadmin:spamreports) allows
people to view spam reports: top 10 reported IPs, top 10 reported users, last
10 reports, reports in last 1 hour, 6 hours, and 24 hours.  Admins can then
take whatever action they think is necessary and delete the report.

This also puts in the functionality to track the IP address of anonymous
comments, whether the journal they're posted in has enabled IP tracking or
not.  These IPs are only stored for 5 days--so anonymous comments deleted
after 5 days are fairly useless, and reports about them are quietly discarded
by the comment deletion system.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@3797 1dfba125-df10-0410-8a77-af1a8d6d074f