New page: /community/create.bml
This page lets you create a community without going through the intermediate
step of creating a personal account.

This patch also changes a bunch of the links to the old page to point to the
new one.

Still to come: removing the old way of doing it, after some policy is worked
out on how to handle people who don't know about the new way of doing things.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@4109 1dfba125-df10-0410-8a77-af1a8d6d074f