Technically, the comment exporter could potentially generate invalid XML if the state wasn't D but it had no body.  This might be the cause of some problems encountered by LJBook?
Fixing that, also adding in a codepath annotation so we can keep track of
how long these pages are taking.  (Seem to be taking a while in my testing.)

And general cleanup, use strict, etc.  Things that are generally good to
have that this file didn't have to begin with.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@4117 1dfba125-df10-0410-8a77-af1a8d6d074f