Add ability for captchas to come from a MogileFS backend.  To do this, set $LJ::CAPTCHA_MOGILEFS = 1.  Also make sure %LJ::MOGILEFS_CONFIG is setup properly--you need a class called 'captcha' in your domain.
This patch also modifies the captchas database table to have a new column,
location, this column has an ENUM for 'blob' or 'mogile' right now, but it
can be expanded in the future.

I also modified the captcha generation jobs to store the generated files in
Mogile if the above options are setup right.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@4451 1dfba125-df10-0410-8a77-af1a8d6d074f