part of the previous fix to captcha.pl from michael and I that I messed up the commit of:
http://changelog.livejournal.com/1735237.html
http://cvs.livejournal.org/browse.cgi/livejournal/bin/maint/captcha.pl.diff?r1=1.13&r2=1.14

Basically, the old captcha gen code would insert into the captcha table
to allocate the capid, but somebody could've got that captcha in the meantime
while the file was still being made.  in practice it didn't happen
often because we keep so many image captchas and they go so quick,
but if it's audio and we're low, or if mogilefs isn't responding like it
wasn't last week, that could cause it.

we're over the problem now, so this is mostly a robustness fix for the
future.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@4601 1dfba125-df10-0410-8a77-af1a8d6d074f