Why yes, Saturday nights are for finding and fixing bugs!
If someone deleted a phone post, it would be gone from userblob because the
data is gone, but the row in phonepostentry remains.  Therefore, later when
they post another phone post, if they had no counter value, it would get
initialized based on the data in userblob.  This would cause the code to
bomb inserting a row into phonepostentry, but it would plow ahead and do
all of the other work, and so the user would see this:

[Access to audio link denied.]

Well, here's the going-forward fix: set the new max for the counter to be
the max of userblob and phonepostentry.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@4670 1dfba125-df10-0410-8a77-af1a8d6d074f