http://zilla.livejournal.org/show_bug.cgi?id=2116
Due to the handling of the (relatively) new memories code, users were unable
to edit/delete memories that pointed to unclustered items.

This update fixes that--users can now delete and edit these memories.  Any
edits will auto-convert the memory to point to the new (journalid, ditemid)
combination instead of just the oldid.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@4682 1dfba125-df10-0410-8a77-af1a8d6d074f