http://zilla.livejournal.org/2205
Fix a bug where an error on the talk form wouldn't select the appropriate
radio button when the form refreshed.

The root cause was that the bottom condition was being triggered and so two
areas were being marked as 'checked' in the HTML and the browser would just
opt to go with the later element to resolve disputes.

This fix requires that a username be specified in order to activate the
third bubble, or, the user must have manually checked it.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@4908 1dfba125-df10-0410-8a77-af1a8d6d074f