make mover paranoid about the set of tables before/after and how it compares tot he set that the mover knows about.
basically we want lisa/matthew to be able to do moves without having
to worry about anything, so the less knowledge about recent code changes
(the more that could be detected automatically) the better.

now the movers can just stay running and they'll just start dying if
there's a problem.


git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@4989 1dfba125-df10-0410-8a77-af1a8d6d074f